+++ CHANGELOG.md
@@ -73,7 +73,14 @@ To be released on Jun 5, 2020.
 [fecf9f]: http://github.com/Gallopsled/pwntools/commit/fecf9f
 [1454]: https://github.com/Gallopsled/pwntools/pull/1454
 
-## 4.1.2 (`stable`)
+## 4.1.3 (`stable`)
+
+- [#1590][1590] Fix `gdb.attach()` for `remote`, `listen`, `ssh` tubes
+  - Also fix `run_in_new_terminal` for Py2 unicode strings
+
+[1590]: https://github.com/Gallopsled/pwntools/pull/1590
+
+## 4.1.2
 
 - Pwntools requires `six` v.1.12.0 or higher